@@ -288,17 +288,17 @@ void LoopClosure::updateGlobalMap() {
   }
 }
 
-void LoopClosure::updatePrimiteEstimatorTrajectory(const nav_msgs::OdometryConstPtr& pose_msg) {
-  // geometry_msgs::PoseStamped pose_stamped;
-  // pose_stamped.header = pose_msg->header;
-  // pose_stamped.header.seq = primitive_estimator_poses_.size() + 1;
-  // pose_stamped.pose = Utility::matrixToRosPose(init_t_w_svin_ * init_t_w_prim_.inverse() *
-  //                                              Utility::rosPoseToMatrix(pose_msg->pose.pose) *
-  //                                              params_.T_body_imu_ * params_.T_imu_cam0_);
-  // primitive_estimator_poses_.push_back(pose_stamped);
-}
+// void LoopClosure::updatePrimiteEstimatorTrajectory(const nav_msgs::OdometryConstPtr& pose_msg) {
+// geometry_msgs::PoseStamped pose_stamped;
+// pose_stamped.header = pose_msg->header;
+// pose_stamped.header.seq = primitive_estimator_poses_.size() + 1;
+// pose_stamped.pose = Utility::matrixToRosPose(init_t_w_svin_ * init_t_w_prim_.inverse() *
+//                                              Utility::rosPoseToMatrix(pose_msg->pose.pose) *
+//                                              params_.T_body_imu_ * params_.T_imu_cam0_);
+// primitive_estimator_poses_.push_back(pose_stamped);
+// }
 
-bool LoopClosure::healthCheck(const TrackingInfo& tracking_info, boost::optional<std::string> error_message) {
+bool LoopClosure::healthCheck(const TrackingInfo& tracking_info, std::string error_message) {
   std::stringstream ss;
   std::setprecision(5);
 
@@ -327,11 +327,9 @@ bool LoopClosure::healthCheck(const TrackingInfo& tracking_info, boost::optional
       static_cast<float>(tracking_info.num_new_keypoints_) / static_cast<float>(tracking_info.num_tracked_keypoints_);
 
   if (new_detected_keypoints_ratio >= 0.75) {
-    if (error_message) {
-      ss << "Too many new keypoints: " << new_detected_keypoints_ratio << std::endl;
-      error_message = ss.str();
-      return false;
-    }
+    ss << "Too many new keypoints: " << new_detected_keypoints_ratio << std::endl;
+    error_message = ss.str();
+    return false;
   }
 
   double average_response = std::accumulate(tracking_info.keypoints_response_strengths_.begin(),
@@ -345,11 +343,9 @@ bool LoopClosure::healthCheck(const TrackingInfo& tracking_info, boost::optional
       static_cast<float>(tracking_info.keypoints_response_strengths_.size());
 
   if (fraction_with_low_detector_response >= 0.85) {
-    if (error_message) {
-      ss << "Too many detectors with low response: " << fraction_with_low_detector_response << std::endl;
-      error_message = ss.str();
-      return false;
-    }
+    ss << "Too many detectors with low response: " << fraction_with_low_detector_response << std::endl;
+    error_message = ss.str();
+    return false;
   }
 
   return true;
