@@ -9,7 +9,7 @@ const Eigen::Vector3d CameraPoseVisualization::lt1 = Eigen::Vector3d(-0.7, -0.2,
 const Eigen::Vector3d CameraPoseVisualization::lt2 = Eigen::Vector3d(-1.0, -0.2, 1.0);
 const Eigen::Vector3d CameraPoseVisualization::oc = Eigen::Vector3d(0.0, 0.0, 0.0);
 
-void Eigen2Point(const Eigen::Vector3d& v, geometry_msgs::Point& p) {  // NOLINT
+void Eigen2Point(const Eigen::Vector3d& v, geometry_msgs::msg::Point& p) {  // NOLINT
   p.x = v.x();
   p.y = v.y();
   p.z = v.z();
@@ -46,19 +46,19 @@ void CameraPoseVisualization::setOpticalCenterConnectorColor(float r, float g, f
 void CameraPoseVisualization::setScale(double s) { m_scale = s; }
 void CameraPoseVisualization::setLineWidth(double width) { m_line_width = width; }
 void CameraPoseVisualization::add_edge(const Eigen::Vector3d& p0, const Eigen::Vector3d& p1) {
-  visualization_msgs::Marker marker;
+  visualization_msgs::msg::Marker marker;
 
   marker.ns = m_marker_ns;
   marker.id = loop_edge_markers.size() + 1;
-  marker.type = visualization_msgs::Marker::LINE_LIST;
-  marker.action = visualization_msgs::Marker::ADD;
+  marker.type = visualization_msgs::msg::Marker::LINE_LIST;
+  marker.action = visualization_msgs::msg::Marker::ADD;
   marker.scale.x = 0.01;
 
   marker.color.b = 1.0f;
   marker.color.a = 1.0;
-  marker.pose = geometry_msgs::Pose();
+  marker.pose = geometry_msgs::msg::Pose();
 
-  geometry_msgs::Point point0, point1;
+  geometry_msgs::msg::Point point0, point1;
 
   Eigen2Point(p0, point0);
   Eigen2Point(p1, point1);
@@ -71,12 +71,12 @@ void CameraPoseVisualization::add_edge(const Eigen::Vector3d& p0, const Eigen::V
 
 void CameraPoseVisualization::add_loopedge(const Eigen::Vector3d& p0, const Eigen::Vector3d& p1) {
   // m_markers.clear();
-  visualization_msgs::Marker marker;
+  visualization_msgs::msg::Marker marker;
 
   marker.ns = m_marker_ns;
   marker.id = loop_edge_markers.size() + 1;
-  marker.type = visualization_msgs::Marker::LINE_STRIP;
-  marker.action = visualization_msgs::Marker::ADD;
+  marker.type = visualization_msgs::msg::Marker::LINE_STRIP;
+  marker.action = visualization_msgs::msg::Marker::ADD;
   marker.scale.x = 0.02;
   marker.color.r = 1.0f;
   marker.color.a = 1.0;
@@ -89,7 +89,7 @@ void CameraPoseVisualization::add_loopedge(const Eigen::Vector3d& p0, const Eige
   marker.pose.orientation.y = 0.0;
   marker.pose.orientation.z = 0.0;
 
-  geometry_msgs::Point point0, point1;
+  geometry_msgs::msg::Point point0, point1;
 
   Eigen2Point(p0, point0);
   Eigen2Point(p1, point1);
@@ -101,12 +101,12 @@ void CameraPoseVisualization::add_loopedge(const Eigen::Vector3d& p0, const Eige
 }
 
 void CameraPoseVisualization::add_pose(const Eigen::Vector3d& p, const Eigen::Quaterniond& q) {
-  visualization_msgs::Marker marker;
+  visualization_msgs::msg::Marker marker;
 
   marker.ns = m_marker_ns;
   marker.id = 0;
-  marker.type = visualization_msgs::Marker::LINE_STRIP;
-  marker.action = visualization_msgs::Marker::ADD;
+  marker.type = visualization_msgs::msg::Marker::LINE_STRIP;
+  marker.action = visualization_msgs::msg::Marker::ADD;
   marker.scale.x = m_line_width;
 
   marker.pose.position.x = 0.0;
@@ -117,7 +117,7 @@ void CameraPoseVisualization::add_pose(const Eigen::Vector3d& p, const Eigen::Qu
   marker.pose.orientation.y = 0.0;
   marker.pose.orientation.z = 0.0;
 
-  geometry_msgs::Point pt_lt, pt_lb, pt_rt, pt_rb, pt_oc, pt_lt0, pt_lt1, pt_lt2;
+  geometry_msgs::msg::Point pt_lt, pt_lb, pt_rt, pt_rb, pt_oc, pt_lt0, pt_lt1, pt_lt2;
 
   Eigen2Point(q * (m_scale * imlt) + p, pt_lt);
   Eigen2Point(q * (m_scale * imlb) + p, pt_lb);
@@ -193,8 +193,9 @@ void CameraPoseVisualization::reset() {
 
 void CameraPoseVisualization::clearCameraPoseMarkers() { camera_pose_markers.clear(); }
 
-void CameraPoseVisualization::publish_by(ros::Publisher& pub, const std_msgs::Header& header) {
-  visualization_msgs::MarkerArray markerArray_msg;
+void CameraPoseVisualization::publish_by(const rclcpp::Publisher<visualization_msgs::msg::MarkerArray>::SharedPtr pub,
+                                         const std_msgs::msg::Header& header) {
+  visualization_msgs::msg::MarkerArray markerArray_msg;
   // int k = (int)m_markers.size();
   /*
   for (int i = 0; i < 5 && k > 0; i++)
@@ -215,14 +216,15 @@ void CameraPoseVisualization::publish_by(ros::Publisher& pub, const std_msgs::He
     markerArray_msg.markers.push_back(marker);
   }
 
-  pub.publish(markerArray_msg);
+  pub->publish(markerArray_msg);
 }
 
-void CameraPoseVisualization::publish_image_by(ros::Publisher& pub, const std_msgs::Header& header) {
-  image.header = header;
+// void CameraPoseVisualization::publish_image_by(rclcpp::Publisher<sensor_msgs::msg::Image>& pub,
+//                                                const std_msgs::msg::Header& header) {
+//   image.header = header;
 
-  pub.publish(image);
-}
+//   pub.publish(image);
+// }
 /*
 void CameraPoseVisualization::add_image(const Eigen::Vector3d& T, const Eigen::Matrix3d& R, const cv::Mat &src)
 {
