@@ -1,15 +1,14 @@
 #include "utils/UtilsOpenCV.h"
 
 #include <cv_bridge/cv_bridge.h>
-#include <ros/ros.h>
-#include <sensor_msgs/Image.h>
+#include <glog/logging.h>
 
 #include <algorithm>
+#include <sensor_msgs/msg/image.hpp>
 #include <string>
 #include <vector>
 
 #include "utils/Utils.h"
-
 /* -------------------------------------------------------------------------- */
 // add circles in the image at desired position/size/color
 void UtilsOpenCV::DrawCirclesInPlace(cv::Mat& img,
@@ -267,14 +266,13 @@ cv::Mat UtilsOpenCV::DrawCornersMatches(const cv::Mat& img1,
   return canvas;
 }
 
-cv::Mat UtilsOpenCV::readRosImage(const sensor_msgs::ImageConstPtr& img_msg, bool grayscale) {
+cv::Mat UtilsOpenCV::readRosImage(const sensor_msgs::msg::Image::ConstSharedPtr img_msg, bool grayscale) {
   cv_bridge::CvImageConstPtr cv_ptr;
   try {
     // TODO(Toni): here we should consider using toCvShare...
     cv_ptr = cv_bridge::toCvCopy(img_msg);
   } catch (cv_bridge::Exception& exception) {
-    ROS_FATAL("cv_bridge exception: %s", exception.what());
-    ros::shutdown();
+    LOG(FATAL) << "cv_bridge exception: " << exception.what();
   }
 
   const cv::Mat img_const = cv_ptr->image;  // Don't modify shared image in ROS.
@@ -294,9 +292,9 @@ cv::Mat UtilsOpenCV::readRosImage(const sensor_msgs::ImageConstPtr& img_msg, boo
     }
     return converted_img;
   } else {
-    ROS_ERROR_STREAM_COND(cv_ptr->encoding != sensor_msgs::image_encodings::MONO8,
-                          "Expected image with MONO8, BGR8, or RGB8 encoding."
-                          "Add in here more conversions if you wish.");
+    LOG_IF(ERROR, cv_ptr->encoding != sensor_msgs::image_encodings::MONO8)
+        << "Expected image with MONO8, BGR8, or RGB8 encoding."
+           "Add in here more conversions if you wish.";
     return img_const;
   }
 }
